// src/shared/api/generated/model/sessionRecordsResponseDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Yoclick API
 * API multi-centro de reservas en marca blanca. Contrato canónico para la app.
 * OpenAPI spec version: 0.1.0
 */
import type { SessionRecordsResponseDtoRecordsItem } from './sessionRecordsResponseDtoRecordsItem.ts';
import type { SessionRecordsResponseDtoTotalsItem } from './sessionRecordsResponseDtoTotalsItem.ts';

export interface SessionRecordsResponseDto {
  timezone: string;
  records: SessionRecordsResponseDtoRecordsItem[];
  totals: SessionRecordsResponseDtoTotalsItem[];
}